Style updates
Made functions start with capital letters. Changed naming style of file names. Switched to pragma once. Placed override attribute on functions that are overriding virtual methods.
