+++ src/FileSystemReceiver/file_system_change_directory.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <string>
+
+class FileSystemChangeDirectory 
+{
+public:
+    virtual ~FileSystemChangeDirectory() {}
+
+    virtual void ChangeDirectory(std::string) = 0;
+};
