+++ src/FileSystemReceiver/file_system_execute_file.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <string>
+
+class FileSystemExecuteFile {
+public:
+    virtual ~FileSystemExecuteFile() {}
+
+    virtual void Execute(std::string) = 0;
+};
