+++ src/FileSystemReceiver/file_system_is_file_at_current_directory.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <string>
+
+class FileSystemIsFileAtCurrentDirectory {
+public:
+    virtual ~FileSystemIsFileAtCurrentDirectory() {}
+
+    virtual bool IsFileAtCurrentDirectory() = 0;
+};
