+++ src/FileSystemReceiver/file_system_is_file_at_given_directory.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <string>
+
+class FileSystemIsFileAtGivenDirectory {
+public:
+    virtual ~FileSystemIsFileAtGivenDirectory() {}
+
+    virtual bool IsFileAtGivenDirectory(std::string) = 0;
+};
