+++ src/FileSystemReceiver/file_system_list_directory.h
@@ -1,0 +1,8 @@
+#pragma once
+
+class FileSystemListDirectory {
+public:
+    virtual ~FileSystemListDirectory() {}
+
+    virtual void ListDirectory() = 0;
+};
